--- app/client/src/ee/constants/errorMessages/5xx.ts
@@ -1,1 +1,0 @@
-export * from "ce/constants/errorMessages/5xx";
